backend/requirements.txt
Django==5.0.2
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.10
python-dotenv==1.0.1
bcrypt==4.1.2
django-filter==23.5
dj-database-url==2.1.0
django-environ==0.11.2 